Separate bot related stuff into "tgbot"
